guvvy/backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# MongoDB & Beanie ODM
beanie==1.22.6
motor==3.3.1

# Utilities
email-validator==2.0.0.post2
httpx==0.25.0
tenacity==8.2.3
aiohttp==3.8.6

# CORS
starlette==0.27.0

# Monitoring
prometheus-fastapi-instrumentator==6.1.0